refactor(cli): extract local sleep/wake lifecycle into reusable lib helpers
