+++ cli/src/commands/sleep.ts
@@ -1,49 +1,6 @@
-import { existsSync, readFileSync } from "fs";
-import { join } from "path";
-
-import {
-  getDaemonPidPath,
-  resolveTargetAssistant,
-} from "../lib/assistant-config.js";
-import type { AssistantEntry } from "../lib/assistant-config.js";
+import { resolveTargetAssistant } from "../lib/assistant-config.js";
 import { dockerResourceNames, sleepContainers } from "../lib/docker.js";
-import { isProcessAlive, stopProcessByPidFile } from "../lib/process";
-
-const ACTIVE_CALL_LEASES_FILE = "active-call-leases.json";
-
-type ActiveCallLease = {
-  callSessionId: string;
-};
-
-function getAssistantRootDir(entry: AssistantEntry): string {
-  if (!entry.resources) {
-    throw new Error(
-      `Local assistant '${entry.assistantId}' is missing resource configuration. Re-hatch to fix.`,
-    );
-  }
-  return join(entry.resources.instanceDir, ".vellum");
-}
-
-function readActiveCallLeases(vellumDir: string): ActiveCallLease[] {
-  const path = join(vellumDir, ACTIVE_CALL_LEASES_FILE);
-  if (!existsSync(path)) {
-    return [];
-  }
-
-  const raw = JSON.parse(readFileSync(path, "utf-8")) as {
-    version?: number;
-    leases?: Array<{ callSessionId?: unknown }>;
-  };
-  if (raw.version !== 1 || !Array.isArray(raw.leases)) {
-    throw new Error(`Invalid active call lease file at ${path}`);
-  }
-
-  return raw.leases.filter(
-    (lease): lease is ActiveCallLease =>
-      typeof lease?.callSessionId === "string" &&
-      lease.callSessionId.length > 0,
-  );
-}
+import { sleepLocalAssistant } from "../lib/local-lifecycle.js";
 
 export async function sleep(): Promise<void> {
   const args = process.argv.slice(3);
@@ -96,65 +53,10 @@ export async function sleep(): Promise<void> {
     process.exit(1);
   }
 
-  if (!entry.resources) {
-    console.error(
-      `Error: Local assistant '${entry.assistantId}' is missing resource configuration. Re-hatch to fix.`,
-    );
+  try {
+    await sleepLocalAssistant(entry, { force });
+  } catch (err) {
+    console.error(`Error: ${err instanceof Error ? err.message : String(err)}`);
     process.exit(1);
-  }
-  const resources = entry.resources;
-  const assistantPidFile = getDaemonPidPath(resources);
-  const vellumDir = getAssistantRootDir(entry);
-  const gatewayPidFile = join(vellumDir, "gateway.pid");
-
-  if (!force) {
-    const assistantAlive = isProcessAlive(assistantPidFile).alive;
-    if (assistantAlive) {
-      try {
-        const activeCallLeases = readActiveCallLeases(vellumDir);
-        if (activeCallLeases.length > 0) {
-          const activeIds = activeCallLeases.map(
-            (lease) => lease.callSessionId,
-          );
-          console.error(
-            `Error: assistant is staying awake for active phone calls (${activeIds.join(
-              ", ",
-            )}). Use 'vellum sleep --force' to stop it anyway.`,
-          );
-          process.exit(1);
-        }
-      } catch (err) {
-        console.error(
-          `Error: ${
-            err instanceof Error ? err.message : String(err)
-          }. Use 'vellum sleep --force' to override if you want to stop the assistant anyway.`,
-        );
-        process.exit(1);
-      }
-    }
-  }
-
-  const assistantStopped = await stopProcessByPidFile(
-    assistantPidFile,
-    "assistant",
-  );
-  if (!assistantStopped) {
-    console.log("Assistant is not running.");
-  } else {
-    console.log("Assistant stopped.");
-  }
-
-  // Stop gateway — use a longer timeout because the gateway has a configurable
-  // drain window (5s) before it exits.
-  const gatewayStopped = await stopProcessByPidFile(
-    gatewayPidFile,
-    "gateway",
-    undefined,
-    7000,
-  );
-  if (!gatewayStopped) {
-    console.log("Gateway is not running.");
-  } else {
-    console.log("Gateway stopped.");
   }
 }
