+++ cli/src/commands/wake.ts
@@ -1,22 +1,6 @@
-import { existsSync, readFileSync, writeFileSync } from "fs";
-import { join } from "path";
-
-import {
-  getDaemonPidPath,
-  resolveTargetAssistant,
-  saveAssistantEntry,
-} from "../lib/assistant-config.js";
+import { resolveTargetAssistant } from "../lib/assistant-config.js";
 import { dockerResourceNames, wakeContainers } from "../lib/docker.js";
-import { seedGuardianTokenFromSiblingEnv } from "../lib/guardian-token.js";
-import { resolveProcessState, stopProcessByPidFile } from "../lib/process";
-import {
-  generateLocalSigningKey,
-  isAssistantWatchModeAvailable,
-  isGatewayWatchModeAvailable,
-  startLocalDaemon,
-  startGateway,
-} from "../lib/local";
-import { maybeStartNgrokTunnel } from "../lib/ngrok";
+import { wakeLocalAssistant } from "../lib/local-lifecycle.js";
 
 export async function wake(): Promise<void> {
   const args = process.argv.slice(3);
@@ -88,163 +72,6 @@ export async function wake(): Promise<void> {
     );
     process.exit(1);
   }
-  const resources = entry.resources;
 
-  const pidFile = getDaemonPidPath(resources);
-
-  let daemonRunning = false;
-  const daemonState = await resolveProcessState(
-    pidFile,
-    resources.daemonPort,
-    "Assistant",
-  );
-  if (daemonState.status === "healthy") {
-    if (watch && isAssistantWatchModeAvailable()) {
-      console.log(
-        `Assistant running (pid ${daemonState.pid}) — restarting in watch mode...`,
-      );
-      await stopProcessByPidFile(pidFile, "assistant");
-    } else {
-      daemonRunning = true;
-      if (watch) {
-        console.log(
-          `Assistant running (pid ${daemonState.pid}) — watch mode not available (no source files). Keeping existing process.`,
-        );
-      } else {
-        console.log(`Assistant already running (pid ${daemonState.pid}).`);
-      }
-    }
-  }
-
-  // Resolve the signing key. The gateway persists its own copy to disk at
-  // <instanceDir>/.vellum/protected/actor-token-signing-key. That on-disk key
-  // is the source of truth because it is what the gateway actually used to sign
-  // existing actor tokens. Prefer it over the lockfile value so that tokens
-  // survive upgrades and any scenario where the two diverge.
-  //
-  // NOTE: Removal of this legacy key path read is blocked on removing all use
-  // of the signing key from the assistant daemon. Until then, the on-disk key
-  // must remain the authoritative source.
-  const legacyKeyPath = join(
-    resources.instanceDir,
-    ".vellum",
-    "protected",
-    "actor-token-signing-key",
-  );
-  let signingKey: string | undefined;
-  if (existsSync(legacyKeyPath)) {
-    try {
-      const raw = readFileSync(legacyKeyPath);
-      if (raw.length === 32) {
-        signingKey = raw.toString("hex");
-      }
-    } catch {
-      // Ignore — fall through to lockfile or generate.
-    }
-  }
-  if (!signingKey) {
-    signingKey = resources.signingKey ?? generateLocalSigningKey();
-  }
-  if (signingKey !== resources.signingKey) {
-    entry.resources = { ...resources, signingKey };
-    saveAssistantEntry(entry);
-  }
-
-  let bootstrapSecret = entry.guardianBootstrapSecret;
-  let bootstrapSecretBackfilled = false;
-  if (!bootstrapSecret) {
-    bootstrapSecret = generateLocalSigningKey();
-    entry.guardianBootstrapSecret = bootstrapSecret;
-    saveAssistantEntry(entry);
-    bootstrapSecretBackfilled = true;
-  }
-
-  if (!daemonRunning) {
-    await startLocalDaemon(watch, resources, { foreground, signingKey });
-  }
-
-  // Start gateway
-  {
-    const vellumDir = join(resources.instanceDir, ".vellum");
-    const gatewayPidFile = join(vellumDir, "gateway.pid");
-    const gatewayState = await resolveProcessState(
-      gatewayPidFile,
-      resources.gatewayPort,
-      "Gateway",
-    );
-    const gatewayAlive = gatewayState.status === "healthy";
-    const needsRestart = bootstrapSecretBackfilled && gatewayAlive;
-    if (needsRestart) {
-      const restartWithWatch = watch && isGatewayWatchModeAvailable();
-      if (restartWithWatch) {
-        console.log(
-          `Gateway running (pid ${gatewayState.pid}) — restarting to apply bootstrap secret...`,
-        );
-      } else {
-        console.log(
-          `Gateway running (pid ${gatewayState.pid}) — restarting without watch mode to apply bootstrap secret...`,
-        );
-      }
-      await stopProcessByPidFile(gatewayPidFile, "gateway");
-      await startGateway(restartWithWatch, resources, {
-        signingKey,
-        bootstrapSecret,
-      });
-    } else if (gatewayAlive) {
-      if (watch && isGatewayWatchModeAvailable()) {
-        console.log(
-          `Gateway running (pid ${gatewayState.pid}) — restarting in watch mode...`,
-        );
-        await stopProcessByPidFile(gatewayPidFile, "gateway");
-        await startGateway(watch, resources, { signingKey, bootstrapSecret });
-      } else {
-        if (watch) {
-          console.log(
-            `Gateway running (pid ${gatewayState.pid}) — watch mode not available (no source files). Keeping existing process.`,
-          );
-        } else {
-          console.log(`Gateway already running (pid ${gatewayState.pid}).`);
-        }
-      }
-    } else {
-      await startGateway(watch, resources, { signingKey, bootstrapSecret });
-    }
-  }
-
-  // Self-heal the guardian token when the current environment's config dir
-  // is missing it. Hatch cross-writes the lockfile across env dirs but the
-  // guardian token is only persisted under the hatch-time env, so a desktop
-  // app built under a different VELLUM_ENVIRONMENT can't find a bearer and
-  // cascades into 401 → auth-rate-limit → 429. A sibling env copy is cheap
-  // and strictly additive.
-  if (seedGuardianTokenFromSiblingEnv(entry.assistantId)) {
-    console.log("   Seeded guardian token from sibling environment.");
-  }
-
-  // Auto-start ngrok if webhook integrations (e.g. Telegram) are configured.
-  const workspaceDir = join(resources.instanceDir, ".vellum", "workspace");
-  const ngrokChild = await maybeStartNgrokTunnel(
-    resources.gatewayPort,
-    workspaceDir,
-  );
-  if (ngrokChild?.pid) {
-    const ngrokPidFile = join(resources.instanceDir, ".vellum", "ngrok.pid");
-    writeFileSync(ngrokPidFile, String(ngrokChild.pid));
-  }
-
-  console.log("Wake complete.");
-
-  if (foreground) {
-    console.log("Running in foreground (Ctrl+C to stop)...\n");
-    // Block forever — the daemon is running with inherited stdio so its
-    // output streams to this terminal. When the user hits Ctrl+C, SIGINT
-    // propagates to the daemon child and both exit.
-    await new Promise<void>((resolve) => {
-      process.on("SIGINT", () => {
-        console.log("\nShutting down...");
-        resolve();
-      });
-      process.on("SIGTERM", () => resolve());
-    });
-  }
+  await wakeLocalAssistant(entry, { watch, foreground });
 }
